+++ CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2024-01-18
+## [v15.15.0] - 2024-01-18
+### Added
+- Added Images in `VillaAnnaHomePage` for better navigation
 
 ### Changed
 - Increased width of drawer in `NavMenu` to 60%.
 - Increased font size of `ListItemButton` to 1.2em.
-- Set padding options in sigle pages
+- Set padding options in single pages
 - Changed Card variants to outlined
 
 ## [v15.14.0] - 2024-01-16